Report storage

+++ microSALT/utils/reporter.py
@@ -9,6 +9,7 @@ import os
 import sys
 import smtplib
 
+from datetime import datetime
 
 from os.path import basename
 from email.mime.text  import MIMEText
@@ -23,8 +24,12 @@ from microSALT.store.lims_fetcher import LIMS_Fetcher
 
 class Reporter():
 
-  def __init__(self, config, log, name = ""):
+  def __init__(self, config, log, name = "", output = ""):
     self.name = name
+    if output == "":
+      self.output = os.getcwd()
+    else:
+      self.output = output
     self.config = config
     self.logger = log
     for k, v in config.items():
@@ -33,12 +38,15 @@ class Reporter():
     self.ticketFinder = LIMS_Fetcher(self.config, self.logger)
     self.attachments = list()
     self.error = False
+    self.dt = datetime.now() 
+    self.now = time.strftime("{}.{}.{}_{}.{}.{}".\
+    format(self.dt.year, self.dt.month, self.dt.day, self.dt.hour, self.dt.minute, self.dt.second))
 
   def report(self, type='default', customer='all'):
+    self.start_web()
     if type == 'json_dump':
       self.gen_json()
-    self.start_web() 
-    if type == 'default':
+    elif type == 'default':
       self.gen_typing()
       self.gen_qc()
       #self.gen_resistence()
@@ -54,8 +62,9 @@ class Reporter():
       raise Exception("Report function recieved invalid format")
     self.kill_flask()
     self.mail()
-    for file in self.attachments:
-      os.remove(file)
+    if self.output == "" or self.output == os.getcwd():
+      for file in self.attachments:
+        os.remove(file)
 
   def gen_STtracker(self, customer="all"):
     self.name ="Sequence Type Update"
@@ -65,21 +74,20 @@ class Reporter():
       self.logger.error("Flask instance currently occupied. Possible rogue process. Retry command")
       self.kill_flask()
       sys.exit(-1)
-    outname = "ST_updates.html"
+    outname = "{}/ST_updates_{}.html".format(self.output, self.now)
     open(outname, 'wb').write(r.content)
     self.attachments.append(outname)
 
   def gen_qc(self):
-    name = self.name
     try:
-      self.ticketFinder.load_lims_project_info(name)
+      self.ticketFinder.load_lims_project_info(self.name)
     except Exception as e:
-      self.logger.error("Project {} does not exist".format(name))
+      self.logger.error("Project {} does not exist".format(self.name))
       self.kill_flask()
       sys.exit(-1)
     try:
-      q = requests.get("http://127.0.0.1:5000/microSALT/{}/qc".format(name), allow_redirects=True)
-      outqc = "{}_QC.html".format(self.ticketFinder.data['Customer_ID_project'])
+      q = requests.get("http://127.0.0.1:5000/microSALT/{}/qc".format(self.name), allow_redirects=True)
+      outqc = "{}/{}_QC_{}.html".format(self.output, self.ticketFinder.data['Customer_ID_project'], self.now)
       open(outqc, 'wb').write(q.content)
       self.attachments.append(outqc)  
     except Exception as e:
@@ -87,21 +95,103 @@ class Reporter():
       self.error = True
  
   def gen_typing(self):
-    name = self.name
     try:
-      self.ticketFinder.load_lims_project_info(name)
+      self.ticketFinder.load_lims_project_info(self.name)
     except Exception as e:
-      self.logger.error("Project {} does not exist".format(name))
+      self.logger.error("Project {} does not exist".format(self.name))
       self.kill_flask()
       sys.exit(-1)
     try:
-      r = requests.get("http://127.0.0.1:5000/microSALT/{}/typing/all".format(name), allow_redirects=True)
-      outtype = "{}_Typing.html".format(self.ticketFinder.data['Customer_ID_project'])
+      r = requests.get("http://127.0.0.1:5000/microSALT/{}/typing/all".format(self.name), allow_redirects=True)
+      outtype = "{}/{}_Typing_{}.html".format(self.output, self.ticketFinder.data['Customer_ID_project'], self.now)
       open(outtype, 'wb').write(r.content)
       self.attachments.append(outtype)
     except Exception as e:
       self.logger.error("Flask instance currently occupied. Possible rogue process. Retry command")
       self.error = True
+
+  def gen_resistence(self):
+    self.ticketFinder.load_lims_project_info(self.name)
+    output = "{}/{}_{}.csv".format(self.output,self.name,self.now)
+    excel = open(output, "w+")
+    sample_info = gen_reportdata(self.name)
+    resdict = dict()
+
+    #Load ALL resistance & genes
+    for s in sample_info['samples']:
+      for r in s.resistances:
+        if not (r.resistance in resdict.keys()) and r.threshold == 'Passed':
+          resdict[r.resistance] =list()
+        if r.threshold == 'Passed' and not r.gene in resdict[r.resistance]:
+          resdict[r.resistance].append(r.gene)
+    for k, v in resdict.items():
+      resdict[k] = sorted(v)
+
+    #Header
+    topline = ",,,"
+    botline = "CG Sample ID,Sample ID,Organism,Sequence Type,Thresholds"
+    for k in sorted(resdict.keys()):
+      genes = [''] * len(resdict[k])
+      topline += ",,{}{}".format(k.replace(',',';'),','.join(genes))
+      botline += ",,{}".format(','.join(sorted(resdict[k])))
+    excel.write("{}\n".format(topline))
+    excel.write("{}\n".format(botline))
+
+    #Individual searches
+    for s in sample_info['samples']:
+      tdict = dict()
+      pref = "{},{},{},{},{}".format(s.CG_ID_sample,s.Customer_ID_sample, s.organism, s.ST_status, s.threshold)
+      #Load single sample
+      for r in s.resistances:
+        if not (r.resistance in tdict.keys()) and r.threshold == 'Passed':
+          tdict[r.resistance] =list()
+        if r.threshold == 'Passed' and not r.gene in tdict[r.resistance]:
+          tdict[r.resistance].append(r.gene)
+      #Compare single sample to all
+      hits = ""
+      for res in sorted(resdict.keys()):
+        if res in tdict.keys():
+          hits += ",1"
+          for gen in sorted(resdict[res]):
+            hits += ","
+            if gen in tdict[res]:
+              hits +="1"
+        else:
+          #Commas eq to res + gen length
+          hits += ",,"
+          pad = [''] * len(resdict[res])
+          hits += ','.join(pad)
+
+      excel.write("{}{}\n".format(pref, hits))
+
+    excel.close()
+    self.attachments.append(output)
+
+  def gen_json(self):
+    report = dict()
+    output = "{}/{}_{}.json".format(self.output, self.name, self.now)
+
+    sample_info = gen_reportdata(self.name)
+    analyses = ['blast_pubmlst', 'quast_assembly', 'blast_resfinder_resistence', 'picard_markduplicate', 'microsalt_samtools_stats']
+    for s in sample_info['samples']:
+      report[s.CG_ID_sample] = dict()
+      for a in analyses:
+        if a == 'blast_resfinder_resistence':
+          report[s.CG_ID_sample][a] = list()
+        else:
+          report[s.CG_ID_sample][a] = dict()
+
+      report[s.CG_ID_sample]['blast_pubmlst'] = {'sequence_type':s.ST_status, 'thresholds':s.threshold}
+      report[s.CG_ID_sample]['quast_assembly'] = {'estimated_genome_length':s.genome_length, 'gc_percentage':float(s.gc_percentage), 'n50':s.n50, 'necessary_contigs':s.contigs}
+
+      for r in s.resistances:
+        if not (r.gene in report[s.CG_ID_sample]['blast_resfinder_resistence']) and r.threshold == 'Passed':
+          report[s.CG_ID_sample]['blast_resfinder_resistence'].append(r.gene)
+
+    #json.dumps(report) #Dumps the json directly
+    with open(output, 'w') as outfile:
+      json.dump(report, outfile)
+    self.attachments.append(output)
 
   def mail(self):
     file_name = self.attachments
@@ -130,91 +220,6 @@ class Reporter():
     self.logger.info("Started webserver on http://127.0.0.1:5000/")
     #Hinders requests before server goes up
     time.sleep(0.05)
-
-  def gen_resistence(self):
-    name = self.name
-    self.ticketFinder.load_lims_project_info(name)
-    excel = open("{}.csv".format(name), "w+")
-    sample_info = gen_reportdata(name)
-    resdict = dict()
-
-
-    #Load ALL resistance & genes
-    for s in sample_info['samples']:
-      for r in s.resistances:
-        if not (r.resistance in resdict.keys()) and r.threshold == 'Passed':
-          resdict[r.resistance] =list()
-        if r.threshold == 'Passed' and not r.gene in resdict[r.resistance]:
-          resdict[r.resistance].append(r.gene)          
-    for k, v in resdict.items():
-      resdict[k] = sorted(v)   
-
-    #Header
-    topline = ",,,"
-    botline = "CG Sample ID,Sample ID,Organism,Sequence Type,Thresholds"
-    for k in sorted(resdict.keys()):
-      genes = [''] * len(resdict[k])
-      topline += ",,{}{}".format(k.replace(',',';'),','.join(genes))     
-      botline += ",,{}".format(','.join(sorted(resdict[k])))
-    excel.write("{}\n".format(topline))
-    excel.write("{}\n".format(botline))
-
-    #Individual searches
-    for s in sample_info['samples']:
-      tdict = dict()
-      pref = "{},{},{},{},{}".format(s.CG_ID_sample,s.Customer_ID_sample, s.organism, s.ST_status, s.threshold)
-      #Load single sample
-      for r in s.resistances:
-        if not (r.resistance in tdict.keys()) and r.threshold == 'Passed':
-          tdict[r.resistance] =list()
-        if r.threshold == 'Passed' and not r.gene in tdict[r.resistance]:
-          tdict[r.resistance].append(r.gene)
-      #Compare single sample to all
-      hits = ""
-      for res in sorted(resdict.keys()):
-        if res in tdict.keys():
-          hits += ",1"
-          for gen in sorted(resdict[res]):
-            hits += ","
-            if gen in tdict[res]:
-              hits +="1"
-        else:
-          #Commas eq to res + gen length
-          hits += ",,"
-          pad = [''] * len(resdict[res])       
-          hits += ','.join(pad) 
- 
-      excel.write("{}{}\n".format(pref, hits))
-       
-    excel.close()
-    inpath = "{}/{}.csv".format(os.getcwd(), name)
-    self.attachments.append(inpath)
-
-  def gen_json(self):
-    report = dict()
-
-    sample_info = gen_reportdata(self.name)
-    analyses = ['blast_pubmlst', 'quast_assembly', 'blast_resfinder_resistence', 'picard_markduplicate', 'microsalt_samtools_stats']
-    for s in sample_info['samples']:
-      report[s.CG_ID_sample] = dict()
-      for a in analyses:
-        if a == 'blast_resfinder_resistence':
-          report[s.CG_ID_sample][a] = list()
-        else:
-          report[s.CG_ID_sample][a] = dict()
-            
-      report[s.CG_ID_sample]['blast_pubmlst'] = {'sequence_type':s.ST_status, 'thresholds':s.threshold}
-      report[s.CG_ID_sample]['quast_assembly'] = {'estimated_genome_length':s.genome_length, 'gc_percentage':float(s.gc_percentage), 'n50':s.n50, 'necessary_contigs':s.contigs}
-
-      for r in s.resistances:
-        if not (r.gene in report[s.CG_ID_sample]['blast_resfinder_resistence']) and r.threshold == 'Passed':
-          report[s.CG_ID_sample]['blast_resfinder_resistence'].append(r.gene)
-
-    #json.dumps(report) #Dumps the json directly
-    inpath = "{}/{}.json".format(os.getcwd(), self.name)
-    with open(inpath, 'w') as outfile: 
-      json.dump(report, outfile)
-    self.attachments.append(inpath)
 
   def kill_flask(self):
     self.server.terminate()
